Botna/net10 and maui toolkit camera

+++ WatchMe/Services/OrchestrationService.cs
@@ -1,5 +1,4 @@
-﻿using WatchMe.Camera;
-using WatchMe.Persistance.CloudProviders;
+﻿using CommunityToolkit.Maui.Views;
 using WatchMe.Persistance.Sqlite;
 using WatchMe.Persistance.Sqlite.Tables;
 using WatchMe.Repository;
@@ -9,54 +8,73 @@ namespace WatchMe.Services
 {
     public interface IOrchestrationService
     {
-        void Initialize(CameraView frontCameraView, CameraView backCameraView);
+        Task Initialize(CameraView backCameraView);
         Task InitiateRecordingProcedure();
         Task StopRecordingProcedure();
     }
 
     public class OrchestrationService : IOrchestrationService
     {
-        private readonly ICloudProviderService _cloudProviderService;
         private readonly IFileSystemService _fileSystemService;
         private readonly INotificationService _notificationService;
         private readonly IVideosRepository _videosRepository;
-        private readonly ICameraWrapper _cameraWrapper;
         private readonly IForegroundServiceDispatcher _serviceDispatcher;
 
         private string _videoTimeStamp;
-        private string _frontVideoFileName;
-        private string _backVideoFileName;
 
-        public OrchestrationService(ICloudProviderService cloudProviderService, IFileSystemService fileSystemService, INotificationService notificationService, IDatabaseInitializer databaseInitializer,
-            IVideosRepository videosRepository, ICameraWrapper cameraWrapper, IForegroundServiceDispatcher serviceDispatcher)
+        private string _backVideoFileName;
+        private CameraView _backCameraView;
+
+        public OrchestrationService(IFileSystemService fileSystemService, INotificationService notificationService, IDatabaseInitializer databaseInitializer,
+            IVideosRepository videosRepository, IForegroundServiceDispatcher serviceDispatcher)
         {
-            _cloudProviderService = cloudProviderService;
             _fileSystemService = fileSystemService;
             _notificationService = notificationService;
             _videosRepository = videosRepository;
-            _cameraWrapper = cameraWrapper;
             _serviceDispatcher = serviceDispatcher;
 
             databaseInitializer.Init();
         }
 
-        public void Initialize(CameraView front, CameraView back)
+        public async Task Initialize(CameraView back)
         {
             _videoTimeStamp = DateTime.UtcNow.ToString("yyyyMMddHHmmssffff");
-            _cameraWrapper.Initialize(front, back);
+            _backVideoFileName = $"Back_{_videoTimeStamp}.mp4";
+            _backCameraView = back;
 
-            _frontVideoFileName = $"Front_{_videoTimeStamp}";
-            _backVideoFileName = $"Back_{_videoTimeStamp}";
+
+            var cameraRequest = await Permissions.RequestAsync<Permissions.Camera>();
+            var microphoneRequest = await Permissions.RequestAsync<Permissions.Microphone>();
+            if (cameraRequest != PermissionStatus.Granted || microphoneRequest != PermissionStatus.Granted)
+            {
+                throw new PermissionException("Camera permission is required to use this feature.");
+            }
+
+            var availableCameras = await _backCameraView.GetAvailableCameras(CancellationToken.None);
+            _backCameraView.SelectedCamera = availableCameras.FirstOrDefault();
+
+            await _backCameraView.StartCameraPreview(CancellationToken.None);
+
+            if (MauiProgram.ISEMULATED)
+            {
+                //clean up videos in our SQLLITE
+                var allVideos = await _videosRepository.GetAllVideosAsync();
+                await _videosRepository.DeleteVideosAsync(allVideos.ToArray());
+
+                //clean up videos in our cache directory
+                var files = Directory.GetFiles(FileSystem.Current.CacheDirectory);
+                var mp4s = files.Where(x => x.Contains("mp4", StringComparison.InvariantCultureIgnoreCase)).ToList();
+                foreach (var mp4 in mp4s)
+                {
+                    File.Delete(mp4);
+                }
+            }
+            //dont start it here yet till we have at max one instance fixed.
+            //_serviceDispatcher.StartVUFS();
         }
 
         public async Task InitiateRecordingProcedure()
         {
-            if (MauiProgram.ISEMULATED)
-            {
-                var allVideos = await _videosRepository.GetAllVideosAsync();
-                await _videosRepository.DeleteVideosAsync(allVideos.ToArray());
-            }
-
             if (!MauiProgram.ISEMULATED)
             {
                 var message = "Andrew just started a WatchMe Routine. Click here to watch along: https://www.youtube.com/watch?v=dQw4w9WgXcQ";
@@ -64,28 +82,27 @@ namespace WatchMe.Services
 
             }
 
-            //await StartRecordingAsync(CameraPosition.Front, _frontVideoFileName);
-            await StartRecordingAsync(CameraPosition.Back, _backVideoFileName);
-
-            _serviceDispatcher.StartVUFS();
+            await StartRecordingAsync(_backCameraView, _backVideoFileName);
         }
 
-        private async Task StartRecordingAsync(CameraPosition position, string filename)
+        private async Task StartRecordingAsync(CameraView camera, string filename)
         {
-            var sizes = _cameraWrapper.GetAvailableResolutions(position);
+            camera.ImageCaptureResolution = camera.SelectedCamera.SupportedResolutions.Last();
+
+            var sizes = camera.SelectedCamera.SupportedResolutions;
             Size sizeToUse;
 
             if (MauiProgram.ISEMULATED)
             {
-                sizeToUse = FindSmallestSize(sizes);
+                sizeToUse = FindSmallestSize(sizes.ToList());
             }
             else
             {
                 sizeToUse = sizes.First(x => x.Width == 1920 && x.Height == 1080);
             }
 
-            var path = _fileSystemService.BuildCacheFileDirectory(filename);
-            await _cameraWrapper.StartRecordingAsync(position, path, sizeToUse);
+            camera.ImageCaptureResolution = sizeToUse;
+            await camera.StartVideoRecording(filename);
 
             await _videosRepository.InsertVideosAsync(new Videos()
             {
@@ -93,6 +110,8 @@ namespace WatchMe.Services
                 VideoState = VideoStates.Recording.ToString(),
                 CreatedAt = DateTime.UtcNow
             });
+
+            _serviceDispatcher.StartVUFS();
         }
 
         private Size FindSmallestSize(List<Size> sizes)
@@ -100,47 +119,25 @@ namespace WatchMe.Services
             return sizes.MinBy(size => size.Width * size.Height);
         }
 
+
         public async Task StopRecordingProcedure()
         {
-            //_videoUploadForegroundService.StopVUFS();
-            //await _cameraWrapper.StopCameraAsync(CameraPosition.Front);
-            await _cameraWrapper.StopCameraAsync(CameraPosition.Back);
+            var backMemStream = await _backCameraView.StopVideoRecording(CancellationToken.None);
 
-            //await backCameraStopTask;
-            var backVideoBytesTask = _fileSystemService.MoveVideoToGallery(_backVideoFileName);
+            if (backMemStream == Stream.Null) return;
 
-            //await frontCameraStopTask;
-            //var frontVideoBytesTask = _fileSystemService.MoveVideoToGallery(_frontVideoFileName);
-
-            await Task.WhenAll(/*frontVideoBytesTask*/ backVideoBytesTask);
-
-            //var frontVideo = await _videosRepository.GetVideosByVideoName(_frontVideoFileName);
-            //frontVideo.TotalBytes = (await frontVideoBytesTask).Count();
-            //frontVideo.VideoState = VideoStates.Finished.ToString();
+            if (backMemStream != Stream.Null)
+            {
+                var videoTimeStamp = DateTime.UtcNow.ToString("yyyyMMddHHmmssffff");
+                await _fileSystemService.MoveVideoToGallery((MemoryStream)backMemStream, "Back_" + videoTimeStamp + ".mp4");
+            }
 
             var backVideo = await _videosRepository.GetVideosByVideoName(_backVideoFileName);
-            backVideo.TotalBytes = (await backVideoBytesTask).Count();
+            backVideo.TotalBytes = backMemStream.Length;
             backVideo.VideoState = VideoStates.Finished.ToString();
 
-            //await _videosRepository.UpdateTotalBytesOfVideo(frontVideo.Id, frontVideo.TotalBytes);
             await _videosRepository.UpdateTotalBytesOfVideo(backVideo.Id, backVideo.TotalBytes);
-            await _videosRepository.UpdateStateOfVideos(VideoStates.Finished,backVideo.Id);
-
-
-            //var recordsUpdated = await _videosRepository.UpdateVideosAsync(frontVideo, backVideo);
-
-            //if (!MauiProgram.ISEMULATED)
-            //{
-            //var frontVideoFileStream = _fileSystemService.GetFileStreamOfFile(_frontVideoFileName);
-            //var frontUploadTask = _cloudProviderService.UploadContentToCloud(frontVideoFileStream, _frontVideoFileName);
-
-            //var backVideoFileStream = _fileSystemService.GetFileStreamOfFile(_backVideoFileName);
-            //var backUploadTask = _cloudProviderService.UploadContentToCloud(backVideoFileStream, _backVideoFileName);
-
-            //await Task.WhenAll();
-            //}
-
-            //var allFiles = await _videosRepository.GetAllVideosAsync();
+            await _videosRepository.UpdateStateOfVideo(backVideo.Id, VideoStates.Finished);
         }
     }
 }
